Addressed feedback. Restructured classes.
